Use explicit destination filename for snapshot
yagmail tries to "guess" MIME type for attachments using filename
extension.
urllib.urlretrieve writes retrieved content to a temporary file with no
extension, so yagmail is unable to determine that attachment is an
image.  By using an explicit destination filename with a ".jpg"
extension, yagmail will use Content-type: image/jpeg for the MIME
object.

Also, by using the yagmail inline type, image will be embedded in the
body of the email's HTML part rather that as an attachment (once yagmail
is updated).  Note that this may be undesirable to some--if they prefer
it as a true attachment, rather than inline.  Might be better to have
this as a configuration option?